Filter branches of all disjunctions before distributing.
We might be able to save lots of work if some branches of the
disjunctions are incompatible with the current substitution. If all
branches of one disjunction are incompatible, we can fail immediately.
This strategy aggressively prunes the search tree but requires duplicate
work as we re-check disjunctions after every distribution.