Delay filtering of disjunctions until after unifications
but filter disjunctions before leaving a fresh. This ordering seems to
work really well when conjunctions are in their ideal ordering, but has
terrible behavior when conjunctions are reordered. This is probably
because checking disjunctions is relatively expensive and in the reverse
ordering they're likely to come before unifications. So if we force
checking of disjunctions before leaving a fresh we'll miss the
opportunity to fail more cheaply on a unification.